[Refactor] ImgDetectionsExtended and Detection parsers (#102)
* Parser v2 example (#74)

* Parsers-V2: Refactor Base Parser (#78)

* Parser name change to ParserGenerator.

* Removed MultiClassificationParser, CompositeMessage and create_multiclassification_message (#81)

* Parsers-V2: Add RegressionParser (#77)

* feat: add prediction message

* feat: add regression message creator

* fix: remove AgeGenderParser import (#86)

* [Refactor] [Bug] ppocr parser to v2, fix bug in ParserGenerator (#88)

* New Keypoint msg.

* KeypointParser V2.

* Parsers-V2: Remove Hand landmark parser and Hand message. (#91)

* Remove Hand landmark parser.

* Remove hand keypoint message.

* feat: update ImgDetectionExtended message

* feat: update detection message creator

* Refactor Segmentation parser. (#94)

* Refactor SCRFD parser. (#95)

* Feat/parsing neural network (#84)

Parsing neural network addition to abstract the need to specify specific parsers.

* Parsers-V2: Refactor Palm detection, SuperAnimal, Lane detection and MLSD (#100)

* KeypointParser as parent class to SuperAnimal.

* Store layer names in MLSD parser.

* [refactor] ppdet, create_detection_msg, DetectionParser

* updated yunet, scfrd utils

---------

+++ depthai_nodes/ml/parsers/classification.py
@@ -115,9 +115,7 @@ class ClassificationParser(BaseParser):
                     f"Expected 1 output layer, got {len(layers)} layers. Please provide the output_layer_name."
                 )
 
-            scores = output.getTensor(self.output_layer_name, dequantize=True).astype(
-                np.float32
-            )
+            scores = output.getTensor(self.output_layer_name, dequantize=True)
             scores = np.array(scores).flatten()
 
             if len(scores) != self.n_classes and self.n_classes != 0:
